requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
blinker==1.9.0
cffi==2.0.0
click==8.3.0
dnspython==2.8.0
Flask==3.1.2
Flask-PyMongo==3.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycparser==2.23
PyJWT==2.10.1
pymongo==4.15.3
python-dotenv==1.1.1
Werkzeug==3.1.3
